requirements.txt
Django==2.2
django-crispy-forms==1.7.2
paho-mqtt==1.4.0
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
